+++ index.js
@@ -1,6 +1,7 @@
 const server = require('./server.js');
 const { conn } = require('./db.js');
-const { api } = require('./config')
+const { api } = require('./config');
+const { addProducts } = require('./aux_functions');
 
 const host = api.host
 const port = api.port
@@ -8,6 +9,8 @@ const port = api.port
 // Syncing all the models at once.
 conn.sync({ force: true }).then(() => {
   server.listen(port, host, () => {
-    console.log(`Server is running at ${host}:${port}`)
+    console.log(`Server is running at ${host}:${port}`);
+    addProducts();
+    console.log("Se cargaron los productos.");
 })
 });
