+++ network/routes.js
@@ -2,7 +2,6 @@ const networkProducts = require('../components/products/network');
 const networkProduct = require('../components/product/network');
 const networkAuth = require('../components/auth/network');
 const networkCategory = require('../components/category/network');
-const networkPreLoad = require('../components/preload/network');
 const networkfilterMetods = require('../components/filterMetods/network');
 const networksearch = require('../components/search/network');
 
@@ -11,7 +10,6 @@ const routes = (server) => {
     server.use('/auth', networkAuth);
     server.use('/category', networkCategory);
     server.use('/product', networkProduct);
-    server.use('/pre_load', networkPreLoad);
     server.use('/filterMetods', networkfilterMetods);
     server.use('/search', networksearch);
 }
